+++ src/TokenEnum.hpp
@@ -1,5 +1,5 @@
 /*
- * TokenEnum.h
+ * TokenEnum.hpp
  *
  * Created on Fri Mar 01 2024
  *
