+++ src/custom_errors.hpp
@@ -1,5 +1,5 @@
 /*
- * custom_errors.h
+ * custom_errors.hpp
  *
  * required :
  *
