+++ src/stuff.hpp
